pyproject.toml
[tool.poetry]
name = "langchain-playground"
version = "0.1.0"
description = ""
authors = ["Hammad-Rehman <124163466+developer-hammad-rehman@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.0"
langchain-google-genai = "^2.0.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.0"
psycopg = "^3.2.2"
psycopg-binary = "^3.2.2"
langchain-ollama = "^0.2.0"
pypdf = "^5.0.0"
bs4 = "^0.0.2"
langchain-postgres = "^0.0.12"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
